+++ BE/src/main/java/ifive/idrop/driver/repository/DriverRepository.java
@@ -1,6 +1,6 @@
 package ifive.idrop.driver.repository;
 
-import ifive.idrop.dto.request.DriverListRequest;
+import ifive.idrop.dto.DriverListRequest;
 import ifive.idrop.entity.*;
 import ifive.idrop.pickup.domain.PickUp;
 import ifive.idrop.pickup.repository.PickUpRepository;
